# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DataHub API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtiene la lista de productos",
                "description": "Muestra una lista de todos los productos.",
                "operationId": "1ce27f90dd8c39cea9520bf45d58c7c2",
                "responses": {
                    "200": {
                        "description": "Lista de productos"
                    },
                    "401": {
                        "description": "Token de acceso inválido o no proporcionado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crea un nuevo producto",
                "description": "Almacena un nuevo producto en la base de datos.",
                "operationId": "acaa0d3fdc74fd4b534e70fc016cf7f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Producto A"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción del Producto A"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Producto A"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 99.99
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Descripción del Producto A"
                                        },
                                        "stock": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token de acceso inválido o no proporcionado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtiene un producto",
                "description": "Muestra un producto específico.",
                "operationId": "249e765b6489191fb8167474db8867f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto obtenido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Producto A"
                                        },
                                        "price": {
                                            "type": "number",
                                            "example": 99.99
                                        },
                                        "stock": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Descripción del Producto A"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-01T00:00:00.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-01T00:00:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token de acceso inválido o no proporcionado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualiza un producto",
                "description": "Actualiza un producto existente en la base de datos.",
                "operationId": "436a5b5d1aff1d5fd341389786161044",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Producto Actualizado"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 149.99
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción actualizada del producto"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado"
                    },
                    "401": {
                        "description": "Token de acceso inválido o no proporcionado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Elimina un producto",
                "description": "Elimina un producto de la base de datos.",
                "operationId": "cc01048629edd9759ba1d7ffc989f667",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado"
                    },
                    "401": {
                        "description": "Token de acceso inválido o no proporcionado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/products/search": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Busca productos por nombre o precio",
                "description": "Busca productos por nombre o precio.",
                "operationId": "e8ceb2cfd0011db14b0d8b0e59c537b2",
                "parameters": [
                    {
                        "name": "data",
                        "in": "query",
                        "description": "Dato de búsqueda",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Producto A"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Productos encontrados"
                    },
                    "400": {
                        "description": "El parámetro data es obligatorio para la búsqueda"
                    },
                    "401": {
                        "description": "Token de acceso inválido o no proporcionado"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}